requirements.txt
Pillow
pyinstaller
questionary
rich 